refactor: rename dcr adslot component file for web
