+++ dotcom-rendering/src/components/DynamicSlowMPU.tsx
@@ -18,7 +18,7 @@ import type {
 	DCRFrontCard,
 	DCRGroupedTrails,
 } from '../types/front';
-import { AdSlot } from './AdSlot';
+import { AdSlot } from './AdSlot.web';
 import { LI } from './Card/components/LI';
 import { UL } from './Card/components/UL';
 import type { Loading } from './CardPicture';
