+++ dotcom-rendering/src/components/FixedMediumSlowXIIMPU.tsx
@@ -7,7 +7,7 @@ import {
 } from '../lib/cardWrappers';
 import { shouldPadWrappableRows } from '../lib/dynamicSlices';
 import type { DCRContainerPalette, DCRFrontCard } from '../types/front';
-import { AdSlot } from './AdSlot';
+import { AdSlot } from './AdSlot.web';
 import { LI } from './Card/components/LI';
 import { UL } from './Card/components/UL';
 import type { Loading } from './CardPicture';
