+++ dotcom-rendering/src/components/FixedSmallSlowVMPU.tsx
@@ -5,7 +5,7 @@ import {
 	CardDefault,
 } from '../lib/cardWrappers';
 import type { DCRContainerPalette, DCRFrontCard } from '../types/front';
-import { AdSlot } from './AdSlot';
+import { AdSlot } from './AdSlot.web';
 import { LI } from './Card/components/LI';
 import { UL } from './Card/components/UL';
 import type { Loading } from './CardPicture';
