+++ dotcom-rendering/src/components/SlotBodyEnd.importable.tsx
@@ -18,7 +18,7 @@ import { type AuthStatus, useAuthStatus } from '../lib/useAuthStatus';
 import { useBraze } from '../lib/useBraze';
 import { useOnce } from '../lib/useOnce';
 import type { TagType } from '../types/tag';
-import { AdSlot } from './AdSlot';
+import { AdSlot } from './AdSlot.web';
 import { canShowBrazeEpic, MaybeBrazeEpic } from './SlotBodyEnd/BrazeEpic';
 import {
 	canShowReaderRevenueEpic,
